renaming theme to not be jsx
