--- {{cookiecutter.repo_name}}/src/theme/theme.jsx
@@ -1,22 +1,0 @@
-import { createMuiTheme } from '@material-ui/core/styles';
-import { red } from '@material-ui/core/colors';
-
-// Create a theme instance.
-const theme = createMuiTheme({
-    palette: {
-        primary: {
-            main: '#556cd6',
-        },
-        secondary: {
-            main: '#19857b',
-        },
-        error: {
-            main: red.A400,
-        },
-        background: {
-            default: '#fff',
-        },
-    },
-});
-
-export default theme;
